Add a buildinfo tool to GRIT.
This lets build systems can ask GRIT what it is going to generate before it generates it, and what upstream dependencies that output will have.

This is a manual merge of changes committed 2009/01/15, 2009/06/04, 2009/06/08, 2009/08/07 and 2009/11/10 to the Google-internal branch of grit.  The original author of the tool was Greg Spencer, with several bugfixes and enhancements subsequently submitted by Vitaly Buka.

BUG=http://code.google.com/p/grit-i18n/issues/detail?id=4

Review URL: https://chromiumcodereview.appspot.com/9963100