Change the default max value for integer policies for ADM and ADMX templates.
If not specified ADM and ADMX templates set the max value for integer policies
to 9999 which is far from useful for policies because those often represent
values that are in milliseconds or bytes so large values are quite common.

Ideally we can specify the bounds for every policy but until then we can at least
provide useful default upper bound of 2*10^9.

BUG=chromium:152660
TEST=writer unit tests pass.
Review URL: https://codereview.chromium.org/10984066